+++ CHANGELOG.md
@@ -3,7 +3,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.5.0 (01/21/18)
-- Add support for Form end points.
+- Add support for Form, Email Template end points.
 
 ## 0.4.0 (12/25/17)
 - Add support for CustomField, CustomRedirect API, EmailClicks, and Opportunity end points.